Large refactor
Trying to come up with the best method for implementing event streaming.

Currently able to receive events from the server but not sure how I want to go about parsing the data, tracking changes, and reporting the changes to the user.

The Firebase class was refactored to be follow closer to the java api and most of the logic was moved into FirebaseClient.